feat: migrate f64 to rust_decimal::Decimal for financial precision
Floating-point rounding errors accumulate across trades when using f64
for prices, quantities, and P&L. Since Bybit already sends values as
strings, rust_decimal::Decimal provides exact decimal arithmetic with
zero-loss parsing at service boundaries.

- Proto fields: double → string for lossless transport (13 fields)
- Domain logic: f64 → Decimal in strategy, risk, order crates
- Database: DOUBLE PRECISION → NUMERIC(20, 8) for exact storage
- Config: max_position_size/max_daily_loss now Decimal
- sqlx: enabled rust_decimal feature for native Decimal ↔ NUMERIC

+++ migrations/001_initial.sql
@@ -3,8 +3,8 @@ CREATE TABLE IF NOT EXISTS signals (
     symbol TEXT NOT NULL,
     side TEXT NOT NULL,
     order_type TEXT NOT NULL,
-    price DOUBLE PRECISION NOT NULL,
-    qty DOUBLE PRECISION NOT NULL,
+    price NUMERIC(20, 8) NOT NULL,
+    qty NUMERIC(20, 8) NOT NULL,
     strategy TEXT NOT NULL DEFAULT 'sma_crossover',
     approved BOOLEAN NOT NULL,
     reject_reason TEXT,
@@ -17,15 +17,15 @@ CREATE TABLE IF NOT EXISTS trades (
     order_id TEXT NOT NULL,
     symbol TEXT NOT NULL,
     side TEXT NOT NULL,
-    price DOUBLE PRECISION NOT NULL,
-    qty DOUBLE PRECISION NOT NULL,
+    price NUMERIC(20, 8) NOT NULL,
+    qty NUMERIC(20, 8) NOT NULL,
     status TEXT NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
 
 CREATE TABLE IF NOT EXISTS positions (
     symbol TEXT PRIMARY KEY,
-    qty DOUBLE PRECISION NOT NULL DEFAULT 0,
-    avg_entry_price DOUBLE PRECISION NOT NULL DEFAULT 0,
+    qty NUMERIC(20, 8) NOT NULL DEFAULT 0,
+    avg_entry_price NUMERIC(20, 8) NOT NULL DEFAULT 0,
     updated_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
